refactor(dashboards): extract overview-tab child component
Move KPI card logic (attribution data fetching, performance summary
cards, trend helpers) from the monolithic marketing-impact parent
into a dedicated overview-tab component. The parent becomes a slim
tab orchestrator managing month picker, focus filter, and tab
selection.

LFXV2-1644
